% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_census_tracts_2000_sf}
\alias{md_census_tracts_2000_sf}
\title{Maryland census tracts (2000 vintage)}
\format{
\subsection{\code{md_census_tracts_2000_sf}}{

A simple feature collection with one row per 2000 tract and 8 columns:
\describe{
\item{geoid}{11-digit state+county+tract FIPS code (CTIDFP00)}
\item{state_fips}{2-digit state FIPS code ("24")}
\item{county_fips}{5-digit state+county FIPS code}
\item{tract_code}{6-digit tract code within county}
\item{tract_name}{Tract name (e.g. "Census Tract 7502")}
\item{land_area}{Land area in square meters (ALAND00)}
\item{water_area}{Water area in square meters (AWATER00)}
\item{geometry}{Multipolygon}
}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}
}
\usage{
md_census_tracts_2000_sf
}
\description{
Decennial 2000 census tract boundaries for Maryland. EPSG:26985,
NAD83 / Maryland (meters). An sf object.
}
\details{
Sourced from U.S. Census Bureau TIGER/Line for vintage 2000,
filtered to state FIPS 24 and reprojected to EPSG:26985.
}
\keyword{datasets}
