% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_pumas_2010_sf}
\alias{md_pumas_2010_sf}
\title{Maryland Public Use Microdata Areas (2010 vintage)}
\format{
\subsection{\code{md_pumas_2010_sf}}{

A simple feature collection with one row per 2010 PUMA and 7 columns:
\describe{
\item{geoid}{7-digit STATE+PUMA code}
\item{state_fips}{2-digit state FIPS code ("24")}
\item{puma_code}{5-digit PUMA code}
\item{puma_name}{PUMA name (e.g. "Washington County--Hagerstown City PUMA")}
\item{land_area}{Land area in square meters}
\item{water_area}{Water area in square meters}
\item{geometry}{Multipolygon}
}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}
}
\usage{
md_pumas_2010_sf
}
\description{
Boundaries of Maryland's 2010-vintage PUMAs. Each PUMA aggregates
census tracts to a minimum population of roughly 100,000. The 2010
PUMA geography was used for ACS microdata releases through 2021.
EPSG:26985, NAD83 / Maryland (meters). An sf object.
}
\details{
Sourced from U.S. Census Bureau TIGER/Line 2019 PUMA file
(2010 vintage), filtered to state FIPS 24 and reprojected to EPSG:26985.
}
\keyword{datasets}
